move corruption tests to own module
